backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
gunicorn==22.0.0
sqlmodel==0.0.19
psycopg2-binary==2.9.9
alembic==1.13.1
argon2-cffi==23.1.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
cryptography==42.0.8
celery==5.4.0
redis==5.0.6
flower==2.0.1
httpx==0.27.0
croniter==2.0.5
pyotp==2.9.0
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-mock==3.14.0
deepdiff==7.0.1
rich==13.7.1
pydantic-settings==2.3.1
